Added test for userservice

+++ src/main/java/com/musthafa/springboot/controllers/UserController.java
@@ -2,6 +2,7 @@ package com.musthafa.springboot.controllers;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,6 +16,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import com.musthafa.springboot.exceptions.UserAlreadyExistsException;
+import com.musthafa.springboot.exceptions.UserNotFoundException;
 import com.musthafa.springboot.models.UserEntity;
 import com.musthafa.springboot.services.UserService;
 
@@ -31,30 +34,43 @@ public class UserController {
 
 	@PostMapping("/v1/users")
 	public ResponseEntity<UserEntity> createUser(@RequestBody UserEntity user) {
-		userService.addUser(user);
-		URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{userId}")
-				.buildAndExpand(user.getUserId()).toUri();
-		return ResponseEntity.created(location).build();
+		int response = userService.addUser(user);
+		if (response == 0)
+			throw new UserAlreadyExistsException("User already exists with userId: " + user.getUserId());
+		else {
+			URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{userId}")
+					.buildAndExpand(user.getUserId()).toUri();
+			return ResponseEntity.created(location).build();
+		}
 	}
 
 	@GetMapping("/v1/users/{userId}")
 	public ResponseEntity<UserEntity> getUser(@PathVariable int userId) {
-		return ResponseEntity.ok(userService.listUserById(userId));
+		Optional<UserEntity> user = userService.listUserById(userId);
+		if (user.isPresent())
+			return ResponseEntity.ok(user.get());
+		else
+			throw new UserNotFoundException("No user exists with userId: " + userId);
 	}
 
 	@PutMapping("/v1/users/{userId}")
 	public ResponseEntity<UserEntity> updateUser(@RequestBody UserEntity user, @PathVariable int userId) {
 		user.setUserId(userId);
-		userService.listUserById(userId);
-		userService.updateUser(user);
-		return new ResponseEntity<UserEntity>(HttpStatus.NO_CONTENT);
+		Optional<UserEntity> exsistingUser = userService.listUserById(userId);
+		if (exsistingUser.isPresent()) {
+			userService.updateUser(user);
+			return new ResponseEntity<UserEntity>(HttpStatus.NO_CONTENT);
+		} else
+			throw new UserNotFoundException("No user exists with userId: " + userId);
 	}
 
 	@DeleteMapping("/v1/users/{userId}")
 	public ResponseEntity<UserEntity> deleteUser(@PathVariable int userId) {
-		userService.listUserById(userId);
-		userService.deleteUserById(userId);
-		return new ResponseEntity<UserEntity>(HttpStatus.OK);
-
+		Optional<UserEntity> exsistingUser = userService.listUserById(userId);
+		if (exsistingUser.isPresent()) {
+			userService.deleteUserById(userId);
+			return new ResponseEntity<UserEntity>(HttpStatus.NO_CONTENT);
+		} else
+			throw new UserNotFoundException("No user exists with userId: " + userId);
 	}
 }
